Stage 1 of exercise on opaque type aliases
- Stage 1 consists of changing `ReductionSet` to an opaque type alias
