+++ exercises/exercise_010_opaque_type_aliases_revised/src/main/scala/org/lunatechlabs/dotty/sudoku/SudokuDetailProcessor.scala
@@ -55,7 +55,7 @@ class SudokuDetailProcessor[DetailType <: SudokoDetailType : UpdateSender] priva
     Behaviors.receiveMessage {
     case Update(cellUpdates, replyTo) if ! fullyReduced =>
       val previousState = state
-      val updatedState = mergeState(state, cellUpdates)
+      val updatedState = state.mergeState(cellUpdates)
       if updatedState == previousState && cellUpdates != cellUpdatesEmpty then
         replyTo ! SudokuDetailUnchanged
         Behaviors.same
@@ -68,9 +68,9 @@ class SudokuDetailProcessor[DetailType <: SudokoDetailType : UpdateSender] priva
           val updateSender = summon[UpdateSender[DetailType]]
           // The following can also be written as:
           // given ActorRef[Response] = replyTo
-          // updateSender.sendUpdate(id, stateChanges(state, transformedUpdatedState))         
-          updateSender.sendUpdate(id, stateChanges(state, transformedUpdatedState))(using replyTo)
-          operational(id, transformedUpdatedState, isFullyReduced(transformedUpdatedState))
+          // updateSender.sendUpdate(id, state.stateChanges(transformedUpdatedState))
+          updateSender.sendUpdate(id, state.stateChanges(transformedUpdatedState))(using replyTo)
+          operational(id, transformedUpdatedState, transformedUpdatedState.isFullyReduced)
 
     case Update(cellUpdates, replyTo) =>
       replyTo ! SudokuDetailUnchanged
@@ -84,23 +84,3 @@ class SudokuDetailProcessor[DetailType <: SudokoDetailType : UpdateSender] priva
       operational(id, InitialDetailState, fullyReduced = false)
 
   }
-
-  private def mergeState(state: ReductionSet, cellUpdates: CellUpdates): ReductionSet =
-      cellUpdates.foldLeft(state) {
-      case (stateTally, (index, updatedCellContent)) =>
-        stateTally.updated(index, stateTally(index) & updatedCellContent)
-    }
-
-  private def stateChanges(state: ReductionSet, updatedState: ReductionSet): CellUpdates =
-    (state zip updatedState).zipWithIndex.foldRight(cellUpdatesEmpty) {
-      case (((previousCellContent, updatedCellContent), index), cellUpdates)
-        if updatedCellContent != previousCellContent =>
-        (index, updatedCellContent) +: cellUpdates
-
-      case (_, cellUpdates) => cellUpdates
-    }
-
-  private def isFullyReduced(state: ReductionSet): Boolean =
-    val allValuesInState = state.flatten
-    allValuesInState == allValuesInState.distinct
-
