+++ exercises/exercise_010_opaque_type_aliases_revised/src/main/scala/org/lunatechlabs/dotty/sudoku/SudokuSolver.scala
@@ -6,8 +6,6 @@ import akka.actor.typed.{ ActorRef, Behavior, SupervisorStrategy }
 
 import scala.concurrent.duration.*
 
-final case class SudokuField(sudoku: Sudoku)
-
 object SudokuSolver:
 
   // SudokuSolver Protocol
